+++ apps/frontend/src/app/chat/[name]/page.tsx
@@ -1,66 +1,23 @@
 "use client";
 
-import { useRouter } from "next/navigation";
-import { useRef, useEffect, useState } from "react";
-import { useSelector, useDispatch } from "react-redux";
-
 import InterviewerProfileSection from "./_components/InterviewerProfileSection";
-import ChatArticle from "./_components/ChatArticle";
-import {
-  selectChatContents,
-  selectChatLimit,
-} from "@/store/redux/features/chat/selector";
-import {
-  initializeChatState,
-  START_CHAT,
-} from "@/store/redux/features/chat/slice";
-import useUserStore from "@/store/useUserStore";
-import { useInterviewerStore } from "@/store/useInterviewerStore";
-import { fetchInterview } from "@/apis/interview";
 import styles from "./page.module.css";
 import Text from "@repo/ui/Text";
+import ChatSection from "./_components/ChatSection";
+import useUserStore from "@/store/useUserStore";
+import { useInterviewerStore } from "@/store/useInterviewerStore";
+import { useRouter } from "next/navigation";
 
 export default function Page() {
-  const chatContents = useSelector(selectChatContents);
-  const dispatch = useDispatch();
-  const chatLimit = useSelector(selectChatLimit);
   const router = useRouter();
+
   const { interviewer } = useInterviewerStore();
   const { user } = useUserStore();
 
-  useEffect(() => {
-    if (!user || !interviewer) return;
-
-    try {
-      (async function init() {
-        const data = await fetchInterview({
-          interviewerId: interviewer?.id,
-          reviewerId: user.id,
-        });
-
-        if (!data?.id) return;
-
-        dispatch({
-          type: START_CHAT,
-          payload: {
-            chatId: data?.id,
-            content: "안녕하세요. 간단히 자기소개 부탁드립니다.",
-          },
-        });
-      })();
-    } catch (err) {}
-    return () => {
-      dispatch(initializeChatState(null));
-    };
-  }, [user, interviewer, dispatch]);
-
-  useEffect(() => {
-    if (chatLimit) {
-      router.push("/result");
-    }
-  }, [router, chatLimit]);
-
-  if (!user || !interviewer) return null;
+  if (!user || !interviewer) {
+    router.back();
+    return null;
+  }
 
   return (
     <div className={styles.container}>
@@ -72,24 +29,10 @@ export default function Page() {
         name={interviewer?.name}
         description={interviewer?.description}
       />
-      {chatContents.map(({ speaker, content, status }) => {
-        return (
-          <ChatArticle key={speaker} type={speaker} status={status}>
-            {speaker === "bot" ? (
-              <>
-                <ChatArticle.Avatar src={interviewer?.imgUrl} />
-                <ChatArticle.Speech text={content} />
-              </>
-            ) : (
-              <>
-                <ChatArticle.Speech text={content} />
-                <ChatArticle.Avatar src={user?.imageSrc} />
-                <ChatArticle.RetryCancelSelector />
-              </>
-            )}
-          </ChatArticle>
-        );
-      })}
+      <ChatSection
+        interviewerImg={interviewer?.imgUrl || ""}
+        userImg={user?.imageSrc || ""}
+      />
     </div>
   );
 }
